Додані javadoc до всіх класів, над якими працював Степано Михайло. Збереження короткої назви посади до БД. Перенесення класу валідації номеру телефона до окремого файлу
