+++ src/main/java/ntukhpi/semit/militaryoblik/javafxutils/validators/TextFieldValidator.java
@@ -6,6 +6,12 @@ import lombok.Setter;
 
 import java.util.regex.Pattern;
 
+
+/**
+ * Клас валідації поля вводу текстових даних
+ *
+ * @author Степанов Михайло
+ */
 @Getter
 @Setter
 @AllArgsConstructor
@@ -18,10 +24,16 @@ public class TextFieldValidator {
     private String textField;
     private String errorMsg;
 
+
+    /**
+     * Валідація тексту за встановленими параметрами
+     *
+     * @throws Exception Повідомлення про причину непроходження валідації
+     */
     public void validate() throws Exception {
         boolean isBlank = textField == null || textField.isBlank();
 
-            if (isBlank && isNecessary)
+        if (isBlank && isNecessary)
             throw new Exception("\"" + fieldName + "\" є обов'язковим полем");
 
         if (!isBlank && regex != null && !regex.matcher(textField).matches())
