+++ src/main/resources/javafxview/ReservistsAll.fxml
@@ -1,9 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import java.lang.*?>
-<?import javafx.geometry.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
 <?import javafx.geometry.Insets?>
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.ComboBox?>
@@ -11,6 +7,7 @@
 <?import javafx.scene.control.TableColumn?>
 <?import javafx.scene.control.TableView?>
 <?import javafx.scene.control.TextField?>
+<?import javafx.scene.control.Tooltip?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.ColumnConstraints?>
 <?import javafx.scene.layout.GridPane?>
@@ -61,7 +58,7 @@
                     <children>
                         <TableView fx:id="reservistsTableView" prefHeight="200.0" prefWidth="200.0" styleClass="table-column" stylesheets="@MOStyles.css" GridPane.columnIndex="1" GridPane.columnSpan="9" GridPane.halignment="CENTER" GridPane.rowIndex="4" GridPane.rowSpan="11" GridPane.valignment="CENTER">
                             <columns>
-                                <TableColumn fx:id="pibColumn" maxWidth="350.0" minWidth="-Infinity" prefWidth="300.0" resizable="false" sortable="false" text="ПІБ" styleClass="column-0"/>
+                                <TableColumn fx:id="pibColumn" maxWidth="350.0" minWidth="-Infinity" prefWidth="300.0" resizable="false" sortable="false" styleClass="column-0" text="ПІБ" />
                                 <TableColumn fx:id="drColumn" maxWidth="-Infinity" minWidth="120.0" prefWidth="150.0" resizable="false" sortable="false" text="ДР" />
                                 <TableColumn fx:id="genderColumn" maxWidth="100.0" minWidth="-Infinity" prefWidth="60.0" resizable="false" sortable="false" text="Стать" />
                                 <TableColumn fx:id="trcColumn" maxWidth="-Infinity" minWidth="250.0" prefWidth="250.0" resizable="false" sortable="false" text="Назва ТРЦ" />
@@ -72,7 +69,10 @@
                             </columns>
                         </TableView>
                         <ComboBox fx:id="tckComboBox" onAction="#tckChanged" prefHeight="55.0" prefWidth="290.0" promptText="-Оберіть ТЦК" GridPane.columnIndex="8" GridPane.columnSpan="2" GridPane.rowIndex="1" />
-                        <ComboBox fx:id="instituteComboBox" centerShape="false" onAction="#instituteChanged" prefHeight="40.0" prefWidth="430.0" promptText="-Оберіть інститут" scaleShape="false" GridPane.columnIndex="1" GridPane.columnSpan="3" GridPane.hgrow="NEVER" GridPane.rowIndex="1" />
+                        <ComboBox fx:id="instituteComboBox" centerShape="false" onAction="#instituteChanged" prefHeight="40.0" prefWidth="430.0" promptText="-Оберіть інститут" scaleShape="false" GridPane.columnIndex="1" GridPane.columnSpan="3" GridPane.hgrow="NEVER" GridPane.rowIndex="1">
+                     <tooltip>
+                        <Tooltip text="Інститут" />
+                     </tooltip></ComboBox>
                         <ComboBox fx:id="cathedraComboBox" onAction="#cathedraChanged" prefHeight="55.0" prefWidth="491.0" promptText="-Оберіть кафедру" GridPane.columnIndex="5" GridPane.columnSpan="2" GridPane.rowIndex="1" />
                         <ComboBox fx:id="typeComboBox" onAction="#typeChanged" prefHeight="55.0" prefWidth="285.0" promptText="-Оберіть категорію звань" GridPane.columnIndex="11" GridPane.columnSpan="2" GridPane.rowIndex="1">
                             <GridPane.margin>
